+++ src/Resources/app/administration/src/main.js
@@ -9,3 +9,9 @@
 import './module/sw-dashboard';
 import './module/sw-extension';
 import './module/sw-settings';
+import './module/sw-settings-shopware-updates';
+import './module/sw-settings-usage-data';
+import './module/sw-first-run-wizard';
+import './module/sw-integration';
+import './module/sw-settings-mailer';
+import './module/sw-settings-store';