+++ src/Resources/app/administration/src/module/sw-first-run-wizard/page/index/index.js
@@ -1,0 +1,15 @@
+/**
+ * digitvision
+ *
+ * @category  digitvision
+ * @package   Shopware\Plugins\DvsnDemoshopFoundation
+ * @copyright (c) 2024 digitvision
+ */
+
+import template from './sw-first-run-wizard.html.twig';
+
+const { Component } = Shopware;
+
+Component.override('sw-first-run-wizard', {
+    template,
+});