+++ src/Resources/app/administration/src/module/sw-integration/index.js
@@ -1,0 +1,9 @@
+/**
+ * digitvision
+ *
+ * @category  digitvision
+ * @package   Shopware\Plugins\DvsnDemoshopFoundation
+ * @copyright (c) 2024 digitvision
+ */
+
+import './page/sw-integration-list';