+++ src/Resources/app/administration/src/module/sw-integration/page/sw-integration-list/index.js
@@ -1,0 +1,15 @@
+/**
+ * digitvision
+ *
+ * @category  digitvision
+ * @package   Shopware\Plugins\DvsnDemoshopFoundation
+ * @copyright (c) 2024 digitvision
+ */
+
+import template from './sw-integration-list.html.twig';
+
+const { Component } = Shopware;
+
+Component.override('sw-integration-list', {
+    template,
+});