+++ src/Resources/app/administration/src/module/sw-settings-mailer/page/sw-settings-mailer/index.js
@@ -1,0 +1,15 @@
+/**
+ * digitvision
+ *
+ * @category  digitvision
+ * @package   Shopware\Plugins\DvsnDemoshopFoundation
+ * @copyright (c) 2024 digitvision
+ */
+
+import template from './sw-settings-mailer.html.twig';
+
+const { Component } = Shopware;
+
+Component.override('sw-settings-mailer', {
+    template,
+});