+++ src/Resources/app/administration/src/module/sw-settings-shopware-updates/page/sw-settings-shopware-updates-wizard/index.js
@@ -1,0 +1,20 @@
+/**
+ * digitvision
+ *
+ * @category  digitvision
+ * @package   Shopware\Plugins\DvsnDemoshopFoundation
+ * @copyright (c) 2024 digitvision
+ */
+
+import template from './sw-settings-shopware-updates-wizard.html.twig';
+
+const { Component } = Shopware;
+
+Component.override('sw-settings-shopware-updates-wizard', {
+    template,
+
+    methods: {
+        createdComponent() {
+        },
+    },
+});