+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 # 1.0.0
 
 - initial release
+
+# 1.0.1
+
+- disabled shopware services within settings
+- removed every app for every disabled setting